requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=4.7,<5.0
psycopg2>=2.8,<3.0
celery>=5.1,<6.0
redis>=3.5,<4.0
drf-yasg>=1.20,<2.0
pytest>=6.2,<7.0
pytest-django>=4.4,<5.0
pytest-cov>=2.11,<3.0
